Set LICENSE in stone
It wasn't explicity set in the beginning, but since it can't work without OpenMW and links into openmw.so as a shared library, then it should match the license as well. There is code in here that is already licensed under GPLv2, so with this update we upgrade it to GPLv3.